Added support for segmented stacks
