+++ CMakeLists.txt
@@ -9,6 +9,7 @@ option(QUANTUM_ENABLE_TESTS "Generate 'tests' target" OFF)
 option(QUANTUM_BOOST_STATIC_LIBS "Link with Boost static libraries." ON)
 option(QUANTUM_BOOST_USE_MULTITHREADED "Use Boost multithreaded libraries." ON)
 option(QUANTUM_BOOST_USE_VALGRIND "Use valgrind headers for Boost." OFF)
+option(QUANTUM_USE_SEGMENTED_STACKS "Use Boost segmented stacks for coroutines." OFF)
 
 if (QUANTUM_INSTALL_ROOT)
     set(CMAKE_INSTALL_PREFIX ${QUANTUM_INSTALL_ROOT})
@@ -78,6 +79,9 @@ endif()
 # Look for Boost
 find_package(Boost 1.61 REQUIRED COMPONENTS context)
 if (Boost_FOUND)
+    if (QUANTUM_USE_SEGMENTED_STACKS)
+        add_definitions(-D__QUANTUM_USE_SEGMENTED_STACKS)
+    endif()
     if (QUANTUM_BOOST_USE_VALGRIND)
         add_definitions(-DBOOST_USE_VALGRIND)
     endif()
